Superb Warfare compiles on Fabric 1.21.1
Zero errors, down from 11921 at the start of the port. This pass closed
the last 761 across 147 files: registries and enum extensions, entities
and vehicles, client render and input, items and data attachments,
blocks and the long tail.

The recurring shape of the work was to find someone else's finished port
instead of doing it: Porting Lib's items module covers most of
IItemExtension, its client_events module has the whole InputEvent family,
and RenderType.translucent() is public where NeoForge only widened
RenderType.TRANSLUCENT.

Where NeoForge has no Fabric counterpart at all -- ItemAbilities,
ExplosionEvent.Detonate, enum extension of Rarity and ArmPose -- the call
sites carry a ponytail: comment naming what was lost and the mixin that
would bring it back.

+++ src/main/java/com/atsuishio/superbwarfare/data/IncrementalTagUpdater.kt
@@ -273,7 +273,7 @@ object IncrementalTagUpdater {
 
             if (it.operation != Operation.REMOVE) {
                 val tag = it.value!!
-                writeByte(tag.id)
+                writeByte(tag.id.toInt())
                 tag.write(ByteBufWrapper(this))
             }
         }
